refactor index.ts server and create consts folder
